+++ streampipes-processors-transformation-jvm/src/main/java/org/apache/streampipes/processors/transformation/jvm/processor/state/labeler/number/NumberLabelerController.java
@@ -75,7 +75,7 @@ public class NumberLabelerController extends StandaloneEventProcessingDeclarer<N
 
     String labelName = getLabelName(extractor);
 
-    List<Integer> numberValues = getNumberValues(extractor);
+    List<Double> numberValues = getNumberValues(extractor);
 
     List<String> labelStrings = getLabelStrings(extractor);
